+++ packages/cs-shared/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [5.1.9]
+### Added
+ - `UBSession.signature` can return a fake signature if `authMock` parameter is true
+
 ## [5.1.7]
 ### Added
  - extended property UBDomain.UBModel.packageJSON added to the extended domain; Property value is parsed model `package.json`